Adding zombie mechanics

+++ tugas_m3/Weapon.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Reflection.Emit;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -10,9 +11,9 @@ namespace tugas_m3
     {
         public string Name { get; set; }
         public int Damage { get; set; }
-        public int FireRate { get; set; }
+        public double FireRate { get; set; }
 
-        public Weapon(string name, int damage, int fireRate)
+        public Weapon(string name, int damage, double fireRate)
         {
             this.Name = name;
             this.Damage = damage;
